requirements.txt
Django==3.1.7
python-decouple==3.4
pytest-django
pytest-cov
pytest-pythonpath